requirements-dev.txt
# Test dependencies for the Breast-Cancer-Prediction project
# Install with: pip install -r requirements-dev.txt

# Core test framework
pytest==8.3.5
pytest-cov==6.1.0
pytest-xdist==3.6.1      # parallel test execution with -n auto

# Code quality
flake8==7.1.1
flake8-bugbear==24.10.31

# Type checking (optional)
mypy==1.11.2
pandas-stubs==2.1.4.240218
